trader-stocktitan/requirements.txt
websocket-client>=1.5.0
openai>=1.3.7
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
webull>=0.2.0